backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
aiofiles==23.2.1
